+++ libft/ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: yrabby <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/02 16:03:51 by yrabby            #+#    #+#             */
-/*   Updated: 2019/04/06 10:21:17 by yrabby           ###   ########.fr       */
+/*   Updated: 2019/04/21 15:01:46 by yrabby           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
 	int n;
